Inpatients_2022
It contains data from inpatients up to Jan 2022. This dataset includes PSYvsCTR and PSYvsPSY comparisons (filter with Comparison column). It also includes multiple outcome calculation types within studies (e.g. means and SD AND remission), which can be filtered with the "primary_outcome_type" variable. RoB 1 is used to assess risk of bias. Effect sizes are not calculated (new metapsy standards need to be implemented in the package).